+++ NEWS.md
@@ -1,3 +1,5 @@
+# preproc.iquizoo (development version)
+
 # preproc.iquizoo 2.6.2
 
 ## Bug Fixes